Init svelte
