--- .eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "extends": ["next/core-web-vitals", "next/typescript"],
-  "rules": {
-    "@typescript-eslint/no-explicit-any": "off"
-  }
-}
