+++ src/app.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    %sveltekit.head%
+  </head>
+  <body class="min-h-screen flex flex-col bg-[#325597] text-white" data-sveltekit-preload-data="hover">
+    %sveltekit.body%
+  </body>
+</html>
