--- src/app/database/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Home() {
-  redirect('https://github.com/nhaar/Waddle-Forever/blob/main/guide/DATABASE.md');
-}
