--- src/app/globals.css
@@ -1,20 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-:root {
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-  }
-}
-
-body {
-}
-
-@layer utilities {
-  .text-balance {
-    text-wrap: balance;
-  }
-}
