--- src/components/Footer.tsx
@@ -1,8 +1,0 @@
-export default function Footer(){
-    return(
-        <footer className="max-w-6xl w-full mx-auto py-4">
-            <hr/>
-            <p className="text-sm text-gray-300">Not affiliated with Disney. Waddle Forever is free and not-for-profit.</p>
-        </footer>
-    )
-}
