+++ src/lib/api.ts
@@ -2,7 +2,7 @@ import { z } from "zod"
 import { ErrorResponse } from "./res"
 
 export function createJsonPost(callback: (body: any) => Promise<Response>) {
-  return async (request: Request) => {
+  return async ({ request }: { request: Request }) => {
     let body
     try {
       body = await request.json()
