+++ src/routes/+page.server.ts
@@ -1,0 +1,5 @@
+import { getCurrentVersion } from "$lib/version";
+
+export async function load() {
+  return { version: (await getCurrentVersion()).version };
+}
