+++ src/routes/api/packages/+server.ts
@@ -1,0 +1,11 @@
+export async function GET() {
+  return Response.json({ 
+    packages: [
+      {
+        setting: 'clothing',
+        name: 'Clothing',
+        size: 640
+      }
+    ]
+  }, { status: 200 })
+}
