+++ src/routes/database/+page.server.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load() {
+  return redirect(303, 'https://github.com/nhaar/Waddle-Forever/blob/main/guide/DATABASE.md');
+}
